+++ app/lib/interface.ts
@@ -66,17 +66,19 @@ export interface bioPage {
   buttonLink: string;
   buttonText: string;
   color: 'blue' | 'orange' | 'green';
-  alert: {
-    name: string;
-    buttonText: string;
-    buttonColor: string;
-    dialogTitle: string;
-    dialogDescription: string;
-    closeText: string;
-    closeLink: string;
-    ctaText: string;
-    ctalink: string;
-  };
+  alert: any;
+}
+
+export interface alertDialog {
+  name: string;
+  buttonText: string;
+  buttonColor: string;
+  dialogTitle: string;
+  dialogDescription: string;
+  closeText: string;
+  closeLink: string;
+  ctaText: string;
+  ctalink: string;
 }
 
 export interface fullCategory {