+++ sanity/schemaTypes/alertDialog.ts
@@ -21,21 +21,7 @@ export default {
       title: 'Button Color',
       type: 'string',
       description: 'Enter the color of the button in HEX format. Example: "#00aaff"',
-      // options: {
-      //   list: [
-      //     {title: 'Blue', value: 'blue'},
-      //     {title: 'Orange', value: 'orange'},
-      //     {title: 'Green', value: 'green'},
-      //   ],
-      // },
     },
-    // {
-    //   name: 'buttonlink',
-    //   title: 'Initial Button Link',
-    //   description:
-    //     'Add the full URL including https:// for external links. If linking to a page on this site, use the relative path. Example (homepage): /',
-    //   type: 'string',
-    // },
     {
       name: 'dialogTitle',
       title: 'Dialog Title',